pyproject.toml
[tool.poetry]
name = "stockmarketdata"
version = "0.1.0"
description = "With this bot, you can access realtime market data from stock exchanges"
authors = ["Nikolaienko Stanislav lettersforstanisalv@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.22.1"
python-decouple = "^3.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
bandit = "^1.7.4"
darglint = "^1.8.1"
isort = "^5.10.1"
mypy = "^0.971"
pre-commit = "^2.20.0"
pydocstyle = "^6.1.1"
pylint = "^2.15.0"
pyupgrade = "^2.37.3"
safety = "^2.1.1"
coverage = "^6.4.4"
pytest-html = "^3.1.1"
pytest-cov = "^3.0.0"
black = "^22.8.0"
finnhub-python = "^2.4.14"

[tool.black]
# https://github.com/psf/black
target-version = ["py310"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 310
line_length = 88
known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true


[build-system]

requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
